refactor: refactor All

See merge request s11-webmobile1-sub2/S11P12B302!171

+++ back-end/koplaydev/src/main/java/com/edu/koplay/dto/ChangeAvatarDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChangeAvatarDTO {
-    private int beforeAvatarIdx;
-    private int afterAvatarIdx;
+    private Long beforeAvatarIdx;
+    private Long afterAvatarIdx;
 
 }
